pyproject.toml
[project]
name = "tempus"
version = "1.0.1"
description = ""
authors = [
    {name = "matheusssi"},
    {name = "1llucaz"}
]
requires-python = ">=3.13"
dependencies = [
    "alembic (>=1.18.4)",
    "pipx (>=1.8.0)",
    "sqlalchemy (<=2.1)",
    "pydantic (>=2.5)",
    "psycopg (>=3.3.3)",
    "python-dotenv (>=1.0)",
    "fastapi (>=0.95.0)",
    "uvicorn[standard] (>=0.21.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "tzdata (>=2025.3,<2026.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
